Simplify API of PropertyLookup
Rely on nullable types and no longer require a default value when
looking up a value. It's easy to use the elvis operator in Kotlin
and simplify using null values. Keep both `valueOf` signatures,
for in case of nullable values, we need to differentiate between
a value not set and a value set to null.